autohiDe mode improved

+++ main.go
@@ -18,7 +18,7 @@ import (
 	"github.com/gotk3/gotk3/gtk"
 )
 
-const version = "0.0.1"
+const version = "0.1.0"
 
 var (
 	appDirs                            []string
@@ -33,19 +33,20 @@ var (
 	widgetAnchor, menuAnchor           gdk.Gravity
 	imgSizeScaled                      int
 	currentWsNum, targetWsNum          int64
+	dockWindow                         *gtk.Window
 )
 
 // Flags
 var cssFileName = flag.String("s", "style.css", "Styling: css file name")
 var targetOutput = flag.String("o", "", "name of Output to display the dock on")
 var displayVersion = flag.Bool("v", false, "display Version information")
-var autohide = flag.Bool("d", false, "auto-hiDe: close window when left or a button clicked")
+var autohide = flag.Bool("d", false, "auto-hiDe: show dock when hotspot hovered, close when left or a button clicked")
 var full = flag.Bool("f", false, "take Full screen width/height")
 var numWS = flag.Int64("w", 8, "number of Workspaces you use")
 var position = flag.String("p", "bottom", "Position: \"bottom\", \"top\" or \"left\"")
-var exclusive = flag.Bool("x", false, "set eXclusive zone: move other windows aside")
+var exclusive = flag.Bool("x", false, "set eXclusive zone: move other windows aside; overrides the \"-l\" argument")
 var imgSize = flag.Int("i", 48, "Icon size")
-var layer = flag.String("l", "top", "Layer \"top\" or \"bottom\"")
+var layer = flag.String("l", "overlay", "Layer \"overlay\", \"top\" or \"bottom\"")
 var launcherCmd = flag.String("c", "nwggrid -p", "Command assigned to the launcher button")
 var alignment = flag.String("a", "center", "Alignment in full width/height: \"start\", \"center\" or \"end\"")
 var marginTop = flag.Int("mt", 0, "Margin Top")
@@ -85,7 +86,7 @@ func buildMainBox(tasks []task, vbox *gtk.Box) {
 
 	// scale icons down when their number increases
 	if *imgSize*6/(len(allItems)) < *imgSize {
-		overflow := (len(allItems) - 8) / 3
+		overflow := (len(allItems) - 6) / 3
 		imgSizeScaled = *imgSize * 6 / (6 + overflow)
 	} else {
 		imgSizeScaled = *imgSize
@@ -191,6 +192,54 @@ func buildMainBox(tasks []task, vbox *gtk.Box) {
 	mainBox.ShowAll()
 }
 
+func setupHotSpot(monitor gdk.Monitor, dockWindow *gtk.Window) gtk.Window {
+	w, h := dockWindow.GetSize()
+
+	win, _ := gtk.WindowNew(gtk.WINDOW_TOPLEVEL)
+
+	layershell.InitForWindow(win)
+	layershell.SetMonitor(win, &monitor)
+
+	box, _ := gtk.BoxNew(gtk.ORIENTATION_HORIZONTAL, 0)
+	win.Add(box)
+
+	win.Connect("enter-notify-event", func() {
+		dockWindow.Hide()
+		dockWindow.Show()
+	})
+
+	if *position == "bottom" || *position == "top" {
+		win.SetSizeRequest(w, 10)
+		if *position == "bottom" {
+			layershell.SetAnchor(win, layershell.LAYER_SHELL_EDGE_BOTTOM, true)
+		} else {
+			layershell.SetAnchor(win, layershell.LAYER_SHELL_EDGE_TOP, true)
+		}
+
+		layershell.SetAnchor(win, layershell.LAYER_SHELL_EDGE_LEFT, *full)
+		layershell.SetAnchor(win, layershell.LAYER_SHELL_EDGE_RIGHT, *full)
+	}
+
+	if *position == "left" {
+		win.SetSizeRequest(10, h)
+		layershell.SetAnchor(win, layershell.LAYER_SHELL_EDGE_LEFT, true)
+
+		layershell.SetAnchor(win, layershell.LAYER_SHELL_EDGE_TOP, *full)
+		layershell.SetAnchor(win, layershell.LAYER_SHELL_EDGE_BOTTOM, *full)
+	}
+
+	layershell.SetLayer(win, layershell.LAYER_SHELL_LAYER_TOP)
+
+	layershell.SetMargin(win, layershell.LAYER_SHELL_EDGE_TOP, *marginTop)
+	layershell.SetMargin(win, layershell.LAYER_SHELL_EDGE_LEFT, *marginLeft)
+	layershell.SetMargin(win, layershell.LAYER_SHELL_EDGE_RIGHT, *marginRight)
+	layershell.SetMargin(win, layershell.LAYER_SHELL_EDGE_BOTTOM, *marginBottom)
+
+	layershell.SetExclusiveZone(win, -1)
+
+	return *win
+}
+
 func main() {
 	flag.Parse()
 
@@ -206,13 +255,14 @@ func main() {
 		for {
 			s := <-signalChan
 			if s == syscall.SIGTERM {
-				fmt.Println("SIGTERM received, bye bye!")
+				println("SIGTERM received, bye bye!")
 				gtk.MainQuit()
 			}
 		}
 	}()
 
-	// We don't want multiple instances. Kill the running instance and exit.
+	// Unless we are in autohide mode, we probably want the same key/mouse binding to turn the dock off.
+	// Kill the running instance and exit.
 	lockFilePath := fmt.Sprintf("%s/nwg-dock.lock", tempDir())
 	lockFile, err := singleinstance.CreateLockFile(lockFilePath)
 	if err != nil {
@@ -220,8 +270,12 @@ func main() {
 		if err == nil {
 			i, err := strconv.Atoi(pid)
 			if err == nil {
-				fmt.Println("Running instance found, sending SIGTERM and exiting...")
-				syscall.Kill(i, syscall.SIGTERM)
+				if !*autohide {
+					println("Running instance found, sending SIGTERM and exiting...")
+					syscall.Kill(i, syscall.SIGTERM)
+				} else {
+					println("Already running")
+				}
 			}
 		}
 		os.Exit(0)
@@ -254,28 +308,31 @@ func main() {
 	} else {
 		fmt.Printf("Using style: %s\n", cssFile)
 		screen, _ := gdk.ScreenGetDefault()
-		gtk.AddProviderForScreen(screen, cssProvider, gtk.STYLE_PROVIDER_PRIORITY_USER)
+		gtk.AddProviderForScreen(screen, cssProvider, gtk.STYLE_PROVIDER_PRIORITY_APPLICATION)
 	}
 
 	win, err := gtk.WindowNew(gtk.WINDOW_TOPLEVEL)
 	if err != nil {
 		log.Fatal("Unable to create window:", err)
 	}
+	dockWindow = win
 
 	layershell.InitForWindow(win)
 
+	var output2mon map[string]*gdk.Monitor
 	if *targetOutput != "" {
-		// We need to assign layershell to a monitor, but we only know the output name!
-		name2mon, err := mapOutputs()
+		// We want to assign layershell to a monitor, but we only know the output name!
+		output2mon, err = mapOutputs()
 		if err == nil {
-			layershell.SetMonitor(win, name2mon[*targetOutput])
+			layershell.SetMonitor(win, output2mon[*targetOutput])
 		} else {
-			fmt.Println(err)
+			println(err)
 		}
 	}
 
 	if *exclusive {
 		layershell.AutoExclusiveZoneEnable(win)
+		*layer = "top"
 	}
 
 	if *position == "bottom" || *position == "top" {
@@ -311,10 +368,20 @@ func main() {
 		menuAnchor = gdk.GDK_GRAVITY_WEST
 	}
 
-	if *layer == "top" {
-		layershell.SetLayer(win, layershell.LAYER_SHELL_LAYER_TOP)
+	if *autohide {
+		// we need to cover the hotspot window, to avoid unwanted .Hide() and .Show() the dockWindow
+		layershell.SetLayer(win, layershell.LAYER_SHELL_LAYER_OVERLAY)
+		layershell.SetExclusiveZone(win, -1)
 	} else {
-		layershell.SetLayer(win, layershell.LAYER_SHELL_LAYER_BOTTOM)
+		// otherwise let's leave users freedom of choice
+		if *layer == "top" {
+			layershell.SetLayer(win, layershell.LAYER_SHELL_LAYER_TOP)
+		} else if *layer == "bottom" {
+			layershell.SetLayer(win, layershell.LAYER_SHELL_LAYER_BOTTOM)
+		} else {
+			layershell.SetLayer(win, layershell.LAYER_SHELL_LAYER_OVERLAY)
+			layershell.SetExclusiveZone(win, -1)
+		}
 	}
 
 	layershell.SetMargin(win, layershell.LAYER_SHELL_EDGE_TOP, *marginTop)
@@ -330,7 +397,8 @@ func main() {
 	win.Connect("leave-notify-event", func() {
 		if *autohide {
 			src, err = glib.TimeoutAdd(uint(1000), func() bool {
-				gtk.MainQuit()
+				win.Hide()
+				src = 0
 				return false
 			})
 		}
@@ -360,18 +428,52 @@ func main() {
 	buildMainBox(tasks, alignmentBox)
 
 	glib.TimeoutAdd(uint(150), func() bool {
-		currentTasks, _ := listTasks()
-		if len(currentTasks) != len(oldTasks) || currentWsNum != oldWsNum || refresh {
-			fmt.Println("refreshing...")
-			buildMainBox(currentTasks, alignmentBox)
-			oldTasks = currentTasks
-			oldWsNum = currentWsNum
-			targetWsNum = currentWsNum
-			refresh = false
+		if win.GetVisible() {
+			currentTasks, _ := listTasks()
+			if len(currentTasks) != len(oldTasks) || currentWsNum != oldWsNum || refresh {
+				println("refreshing...")
+				buildMainBox(currentTasks, alignmentBox)
+				oldTasks = currentTasks
+				oldWsNum = currentWsNum
+				targetWsNum = currentWsNum
+				refresh = false
+			}
 		}
 		return true
 	})
 
 	win.ShowAll()
+
+	if *autohide {
+		win.Hide()
+
+		mRefProvider, _ := gtk.CssProviderNew()
+		if err := mRefProvider.LoadFromPath("/usr/share/nwg-dock/hotspot.css"); err != nil {
+			println(err)
+		}
+
+		if *targetOutput == "" {
+			// hot spots on all displays
+			monitors, _ := listMonitors()
+			for _, monitor := range monitors {
+				win := setupHotSpot(monitor, win)
+
+				context, _ := win.GetStyleContext()
+				context.AddProvider(mRefProvider, gtk.STYLE_PROVIDER_PRIORITY_APPLICATION)
+
+				win.ShowAll()
+			}
+		} else {
+			// hot spot on the selected display only
+			monitor, _ := output2mon[*targetOutput]
+			win := setupHotSpot(*monitor, win)
+
+			context, _ := win.GetStyleContext()
+			context.AddProvider(mRefProvider, gtk.STYLE_PROVIDER_PRIORITY_APPLICATION)
+
+			win.ShowAll()
+		}
+	}
+
 	gtk.Main()
 }
